+++ x/observer/types/query.pb.gw.go
@@ -499,7 +499,7 @@ func request_Query_BlameByChainAndNonce_0(ctx context.Context, marshaler runtime
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "chain_id")
 	}
 
-	protoReq.ChainId, err = runtime.Uint32(val)
+	protoReq.ChainId, err = runtime.Int64(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "chain_id", err)
@@ -510,7 +510,7 @@ func request_Query_BlameByChainAndNonce_0(ctx context.Context, marshaler runtime
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "nonce")
 	}
 
-	protoReq.Nonce, err = runtime.Uint64(val)
+	protoReq.Nonce, err = runtime.Int64(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "nonce", err)
@@ -537,7 +537,7 @@ func local_request_Query_BlameByChainAndNonce_0(ctx context.Context, marshaler r
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "chain_id")
 	}
 
-	protoReq.ChainId, err = runtime.Uint32(val)
+	protoReq.ChainId, err = runtime.Int64(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "chain_id", err)
@@ -548,7 +548,7 @@ func local_request_Query_BlameByChainAndNonce_0(ctx context.Context, marshaler r
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "nonce")
 	}
 
-	protoReq.Nonce, err = runtime.Uint64(val)
+	protoReq.Nonce, err = runtime.Int64(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "nonce", err)
